requirements.txt
pyyaml=3.12
latex>=0.6.2.dev1
